refactor(user-profile): renamed to new folder
